chore(field): introduction test suites (for InteractionStateMixin)
