+++ packages/field/package.json
@@ -28,6 +28,7 @@
     "src",
     "stories",
     "test",
+    "test-suites",
     "translations",
     "*.js"
   ],
